+++ src/app/my/(account)/@info/(user-info)/_components/UserInfo/UserInfoSection.tsx
@@ -7,13 +7,13 @@ import { useRouter } from "next/navigation";
 import { UserInfoFormValues } from "@/components/user/UserInfoForm/UserInfoFormValues";
 
 // hooks
-import useUpdateUserInfo from "../../../../../hooks/user/useUpdateUserInfo";
-import useGetUserInfo from "../../../../../hooks/user/useGetUserInfo";
+import useUpdateUserInfo from "../../../../../../../hooks/user/useUpdateUserInfo";
+import useGetUserInfo from "../../../../../../../hooks/user/useGetUserInfo";
 import useModal from "@/hooks/useModal";
 
 // _components
 import UserInfoContent from "./UserInfoContent";
-import EditUserForm from "../EditUserForm/EditUserForm";
+
 import Modal from "@/components/common/Modal/Modal";
 
 import { filterItemsByIds } from "@/utils/filterItemsByIds";
@@ -21,13 +21,14 @@ import { filterItemsByIds } from "@/utils/filterItemsByIds";
 import STACK_LIST from "@/constant/stackList";
 import UserInfoFallback from "../ui/UserInfoFallback";
 import { useToast } from "@/hooks/useToast";
+import EditUserForm from "../EditUserForm/EditUserForm";
 
 const UserInfoSection = () => {
-  const router = useRouter();
   const { toast } = useToast();
   const { modal: edit, openModal, closeModal } = useModal();
 
-  const { userInfo, isError, isFetching } = useGetUserInfo();
+  const { userInfo, isFetching } = useGetUserInfo();
+
   const { mutate, isSuccess } = useUpdateUserInfo();
 
   useEffect(() => {
@@ -68,48 +69,28 @@ const UserInfoSection = () => {
       )}
 
       {/* 유저정보 */}
-      <section className="mb-4">
-        <div className="bg-white rounded-lg p-4 mb-4 shadow-sm">
-          <h2 className=" text-xl text-primary">회원정보</h2>
+
+      {/* 패칭 중일 때 */}
+      {isFetching && <UserInfoFallback />}
+
+      {/* 패칭이 끝난 후 데이터가 존재 했을 때 */}
+      {!isFetching && userInfo && (
+        <div className="bg-white rounded-lg shadow-sm p-6 relative">
+          <UserInfoContent
+            name={userInfo.name || ""}
+            introduce={userInfo.introduce || ""}
+            waringCnt={userInfo.waringCnt || 0}
+            stacks={filterItemsByIds(userInfo.stacks, STACK_LIST) || []}
+          />
+
+          <button
+            onClick={openModal}
+            className="absolute right-4 bottom-4 bg-primary text-sm text-white py-2 px-4 rounded-md hover:bg-opacity-90 transition-colors disabled:bg-opacity-90"
+          >
+            수정하기
+          </button>
         </div>
-
-        {/* 패칭 중일 때 */}
-        {isFetching && <UserInfoFallback />}
-
-        {/* 패칭이 끝난 후 데이터가 존재 했을 때 */}
-        {!isError && !isFetching && userInfo && (
-          <div className="bg-white rounded-lg shadow-sm p-6 relative">
-            <UserInfoContent
-              name={userInfo.name || ""}
-              introduce={userInfo.introduce || ""}
-              waringCnt={userInfo.waringCnt || 0}
-              stacks={filterItemsByIds(userInfo.stacks, STACK_LIST) || []}
-            />
-
-            <button
-              onClick={openModal}
-              className="absolute right-4 bottom-4 bg-primary text-sm text-white py-2 px-4 rounded-md hover:bg-opacity-90 transition-colors disabled:bg-opacity-90"
-            >
-              수정하기
-            </button>
-          </div>
-        )}
-
-        {/* 유저 정보를 불러오지 못했을 때 */}
-        {isError && (
-          <div className="bg-white rounded-lg shadow-sm p-6 relative">
-            <p className="mb-2 text-center">
-              정보를 불러오지 못했습니다. 잠시 후 다시 시도해주세요.
-            </p>
-            <button
-              onClick={() => router.replace("/")}
-              className="w-full py-2 bg-primary text-white rounded-lg"
-            >
-              처음으로 돌아가기
-            </button>
-          </div>
-        )}
-      </section>
+      )}
     </>
   );
 };
