+++ src/app/my/(account)/@info/(user-info)/_components/ui/UserInfoFallback.tsx
@@ -3,7 +3,7 @@ import React from "react";
 
 const UserInfoFallback = () => {
   return (
-    <div className="flex justify-center items-center h-[243px] w-[921px] rounded-lg mb-4 bg-white">
+    <div className="flex items-center justify-center h-[243px] w-full rounded-lg mb-4 bg-white">
       <LoadingIndicator />
     </div>
   );
