+++ src/app/my/(account)/@info/(user-info)/page.tsx
@@ -1,0 +1,7 @@
+import UserInfoContainer from "./_components/UserInfo/UserInfoContainer";
+
+const page = () => {
+  return <UserInfoContainer />;
+};
+
+export default page;
