+++ src/app/my/(account)/@review/(user-review)/page.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import UserReviewContainer from "./_components/UserReview/UserReviewContainer";
+
+const page = () => {
+  return <UserReviewContainer />;
+};
+
+export default page;
